refactor: keyboard linux JBR-6227
