+++ native/JavaKeyCode2X11.cpp
@@ -4,198 +4,198 @@
 #include <X11/keysym.h>
 
 int JavaKeyCode2X11(JNIEnv* env,
-                    ScopedJNIClass* cls /*KeyEvent*/,
+                    jclass cls /*KeyEvent*/,
                     int keycode) {
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_ENTER, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_BACK_SPACE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_TAB, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_CANCEL, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_CLEAR, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_SHIFT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_CONTROL, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_ALT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_PAUSE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_CAPS_LOCK, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_ESCAPE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_SPACE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_PAGE_UP, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_PAGE_DOWN, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_END, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_HOME, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_LEFT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_UP, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_RIGHT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DOWN, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_COMMA, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_MINUS, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_PERIOD, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_SLASH, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_0, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_1, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_2, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_3, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_4, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_5, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_6, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_7, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_8, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_9, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_SEMICOLON, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_EQUALS, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_A, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_B, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_C, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_D, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_E, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_G, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_H, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_I, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_J, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_K, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_L, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_M, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_N, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_O, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_P, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_Q, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_R, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_S, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_T, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_U, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_V, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_W, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_X, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_Y, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_Z, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_OPEN_BRACKET, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_BACK_SLASH, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_CLOSE_BRACKET, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_NUMPAD0, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_NUMPAD1, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_NUMPAD2, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_NUMPAD3, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_NUMPAD4, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_NUMPAD5, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_NUMPAD6, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_NUMPAD7, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_NUMPAD8, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_NUMPAD9, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_MULTIPLY, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_ADD, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_SEPARATER, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_SEPARATOR, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_SUBTRACT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DECIMAL, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DIVIDE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DELETE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_NUM_LOCK, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_SCROLL_LOCK, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F1, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F2, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F3, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F4, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F5, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F6, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F7, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F8, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F9, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F10, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F11, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F12, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F13, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F14, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F15, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F16, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F17, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F18, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F19, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F20, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F21, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F22, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F23, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_F24, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_PRINTSCREEN, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_INSERT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_HELP, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_META, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_BACK_QUOTE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_QUOTE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_KP_UP, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_KP_DOWN, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_KP_LEFT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_KP_RIGHT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_GRAVE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_ACUTE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_CIRCUMFLEX, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_TILDE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_MACRON, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_BREVE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_ABOVEDOT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_DIAERESIS, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_ABOVERING, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_DOUBLEACUTE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_CARON, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_CEDILLA, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_OGONEK, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_IOTA, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_VOICED_SOUND, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DEAD_SEMIVOICED_SOUND, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_AMPERSAND, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_ASTERISK, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_QUOTEDBL, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_LESS, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_GREATER, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_BRACELEFT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_BRACERIGHT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_AT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_COLON, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_CIRCUMFLEX, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_DOLLAR, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_EURO_SIGN, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_EXCLAMATION_MARK, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_INVERTED_EXCLAMATION_MARK,
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_ENTER, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_BACK_SPACE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_TAB, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_CANCEL, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_CLEAR, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_SHIFT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_CONTROL, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_ALT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_PAUSE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_CAPS_LOCK, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_ESCAPE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_SPACE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_PAGE_UP, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_PAGE_DOWN, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_END, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_HOME, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_LEFT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_UP, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_RIGHT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DOWN, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_COMMA, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_MINUS, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_PERIOD, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_SLASH, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_0, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_1, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_2, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_3, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_4, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_5, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_6, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_7, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_8, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_9, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_SEMICOLON, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_EQUALS, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_A, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_B, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_C, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_D, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_E, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_G, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_H, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_I, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_J, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_K, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_L, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_M, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_N, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_O, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_P, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_Q, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_R, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_S, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_T, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_U, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_V, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_W, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_X, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_Y, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_Z, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_OPEN_BRACKET, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_BACK_SLASH, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_CLOSE_BRACKET, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_NUMPAD0, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_NUMPAD1, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_NUMPAD2, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_NUMPAD3, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_NUMPAD4, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_NUMPAD5, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_NUMPAD6, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_NUMPAD7, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_NUMPAD8, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_NUMPAD9, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_MULTIPLY, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_ADD, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_SEPARATER, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_SEPARATOR, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_SUBTRACT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DECIMAL, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DIVIDE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DELETE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_NUM_LOCK, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_SCROLL_LOCK, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F1, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F2, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F3, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F4, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F5, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F6, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F7, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F8, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F9, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F10, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F11, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F12, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F13, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F14, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F15, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F16, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F17, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F18, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F19, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F20, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F21, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F22, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F23, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_F24, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_PRINTSCREEN, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_INSERT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_HELP, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_META, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_BACK_QUOTE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_QUOTE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_KP_UP, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_KP_DOWN, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_KP_LEFT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_KP_RIGHT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_GRAVE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_ACUTE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_CIRCUMFLEX, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_TILDE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_MACRON, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_BREVE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_ABOVEDOT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_DIAERESIS, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_ABOVERING, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_DOUBLEACUTE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_CARON, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_CEDILLA, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_OGONEK, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_IOTA, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_VOICED_SOUND, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DEAD_SEMIVOICED_SOUND, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_AMPERSAND, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_ASTERISK, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_QUOTEDBL, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_LESS, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_GREATER, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_BRACELEFT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_BRACERIGHT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_AT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_COLON, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_CIRCUMFLEX, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_DOLLAR, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_EURO_SIGN, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_EXCLAMATION_MARK, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_INVERTED_EXCLAMATION_MARK,
                            XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_LEFT_PARENTHESIS, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_NUMBER_SIGN, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_PLUS, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_RIGHT_PARENTHESIS, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_UNDERSCORE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_WINDOWS, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_CONTEXT_MENU, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_FINAL, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_CONVERT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_NONCONVERT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_ACCEPT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_MODECHANGE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_KANA, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_KANJI, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_ALPHANUMERIC, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_KATAKANA, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_HIRAGANA, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_FULL_WIDTH, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_HALF_WIDTH, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_ROMAN_CHARACTERS, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_ALL_CANDIDATES, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_PREVIOUS_CANDIDATE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_CODE_INPUT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_JAPANESE_KATAKANA, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_JAPANESE_HIRAGANA, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_JAPANESE_ROMAN, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_KANA_LOCK, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_INPUT_METHOD_ON_OFF, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_CUT, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_COPY, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_PASTE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_UNDO, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_AGAIN, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_FIND, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_PROPS, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_STOP, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_COMPOSE, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_ALT_GRAPH, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_BEGIN, XK_VoidSymbol);
-  JNI_STATIC_DEFINE_INT_RV(env, *cls, VK_UNDEFINED, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_LEFT_PARENTHESIS, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_NUMBER_SIGN, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_PLUS, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_RIGHT_PARENTHESIS, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_UNDERSCORE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_WINDOWS, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_CONTEXT_MENU, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_FINAL, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_CONVERT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_NONCONVERT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_ACCEPT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_MODECHANGE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_KANA, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_KANJI, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_ALPHANUMERIC, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_KATAKANA, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_HIRAGANA, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_FULL_WIDTH, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_HALF_WIDTH, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_ROMAN_CHARACTERS, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_ALL_CANDIDATES, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_PREVIOUS_CANDIDATE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_CODE_INPUT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_JAPANESE_KATAKANA, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_JAPANESE_HIRAGANA, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_JAPANESE_ROMAN, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_KANA_LOCK, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_INPUT_METHOD_ON_OFF, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_CUT, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_COPY, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_PASTE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_UNDO, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_AGAIN, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_FIND, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_PROPS, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_STOP, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_COMPOSE, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_ALT_GRAPH, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_BEGIN, XK_VoidSymbol);
+  JNI_STATIC_DEFINE_INT_RV(env, cls, VK_UNDEFINED, XK_VoidSymbol);
 
   if (keycode >= JNI_STATIC(VK_A) && keycode <= JNI_STATIC(VK_Z))
     return XK_a + (keycode - JNI_STATIC(VK_A));
